app/requirements/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.2
idna==3.4
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.2
python-dotenv==0.18.0
requests==2.31.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
